Kernel Functions Header Changes
- Added Kernel.h Header
- Moved relevant Kernel definitions to Kernel.h
