+++ grass/grass.c
@@ -10,8 +10,7 @@
  */
 
 #include "egos.h"
-#include "syscall.h"
-#include "process.h"
+#include "kernel.h"
 
 struct grass *grass = (void*)APPS_STACK_TOP;
 struct earth *earth = (void*)GRASS_STACK_TOP;
