+++ grass/kernel.c
@@ -12,9 +12,15 @@
 
 
 #include "egos.h"
-#include "syscall.h"
-#include "process.h"
+#include "kernel.h"
 #include <string.h>
+
+uint proc_curr_idx;
+struct process proc_set[MAX_NPROCESS];
+struct pending_ipc *msg_buffer;
+
+static void proc_yield();
+static void proc_try_syscall(struct process *proc);
 
 void kernel_entry(uint is_interrupt, uint id) {
     /* Save process context */
@@ -28,9 +34,6 @@ void kernel_entry(uint is_interrupt, uint id) {
     memcpy(SAVED_REGISTER_ADDR, proc_set[proc_curr_idx].saved_register, SAVED_REGISTER_SIZE);
 }
 
-#define EXCP_ID_ECALL_U    8
-#define EXCP_ID_ECALL_M    11
-
 void excp_entry(uint id) {
     /* Student's code goes here (system call and memory exception). */
 
@@ -41,16 +44,6 @@ void excp_entry(uint id) {
     /* Student's code ends here. */
     FATAL("excp_entry: kernel got exception %d", id);
 }
-
-#define INTR_ID_SOFT       3
-#define INTR_ID_TIMER      7
-
-static void proc_yield();
-static void proc_try_syscall(struct process *proc);
-
-uint proc_curr_idx;
-struct process proc_set[MAX_NPROCESS];
-struct pending_ipc *msg_buffer;
 
 void intr_entry(uint id) {
     if (id == INTR_ID_TIMER && curr_pid < GPID_SHELL) {
@@ -118,7 +111,7 @@ static int proc_try_send(struct syscall *sc, struct process *sender) {
     
     for (uint i = 0; i < MAX_NPROCESS; i++) {
         struct process dst = proc_set[i];
-        if (dst.pid == sc->msg.receiver) {
+        if (dst.pid == sc->msg.receiver && dst.status != PROC_UNUSED) {
             /* Destination is not receiving, or will not take msg from sender */
             if (! (dst.status == PROC_PENDING && dst.pending_syscall == SYS_RECV) )   return -1;
             if (! (dst.receive_from == GPID_ALL || dst.receive_from == sender->pid) ) return -1;
